requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
python-dotenv>=1.0
requests>=2.31
gunicorn>=21.0
psycopg2-binary>=2.9.0
